Change method from util to UserContext
