+++ app/src/main/java/corp/kairos/adamastor/UserContext.java
@@ -54,5 +54,26 @@ public class UserContext {
         this.location = location;
     }
 
+
+
+    public static String getContextNameById(int id) {
+        switch (id) {
+            case 0:
+                return "Home";
+
+            case 1:
+                return "Work";
+
+            case 2:
+                return "Travel";
+
+            case 3:
+                return "Fitness";
+
+            default:
+                return "Other";
+        }
+    }
+
 }
