+++ app/src/main/res/layout/statistics_app_layout.xml
@@ -57,7 +57,7 @@
         android:id="@+id/app_stats_progress_bar"
         style="@android:style/Widget.Material.ProgressBar.Horizontal"
         android:layout_width="160dp"
-        android:layout_height="30dp"
+        android:layout_height="36dp"
         android:layout_alignParentLeft="true"
         android:layout_alignTop="@+id/app_stats_label"
         android:layout_marginStart="8dp"
